requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-django==4.12.0
sqlparse==0.5.5
